Switch to FreeGlut
Glut does not have any means of safely existing the MainLoop. I used `exit(0);` but that made `nvapi.dll` to throw an exception and I didn't have `nvapi.pdb` to have the symbols from that dll and was not able to understand the reason.
After learning that FreeGlut has a `glutLeaveMainLoop` function used that to exit the MainLoop and called `ovr_Destroy` and `_Shutdown` outside of the loop.
Which got rid off the thrown exception.